Feature/dsnd 1423

+++ src/main/java/uk/gov/companieshouse/pscstatementdataapi/services/PscStatementService.java
@@ -3,14 +3,17 @@ package uk.gov.companieshouse.pscstatementdataapi.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import uk.gov.companieshouse.api.InternalApiClient;
 import uk.gov.companieshouse.api.metrics.MetricsApi;
+import uk.gov.companieshouse.api.model.ApiResponse;
 import uk.gov.companieshouse.api.psc.CompanyPscStatement;
 import uk.gov.companieshouse.api.psc.Statement;
 import uk.gov.companieshouse.api.psc.StatementLinksType;
 import uk.gov.companieshouse.api.psc.StatementList;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.pscstatementdataapi.api.CompanyMetricsApiService;
+import uk.gov.companieshouse.pscstatementdataapi.api.PscStatementApiService;
 import uk.gov.companieshouse.pscstatementdataapi.exception.BadRequestException;
 import uk.gov.companieshouse.pscstatementdataapi.exception.ResourceNotFoundException;
 import uk.gov.companieshouse.pscstatementdataapi.model.Created;
@@ -40,6 +43,8 @@ public class PscStatementService {
 
   @Autowired
   InternalApiClient internalApiClient;
+  @Autowired
+  PscStatementApiService apiClientService;
 
   public Statement retrievePscStatementFromDb(String companyNumber, String statementId) throws JsonProcessingException, ResourceNotFoundException {
     PscStatementDocument pscStatementDocument = getPscStatementDocument(companyNumber, statementId);
@@ -60,9 +65,23 @@ public class PscStatementService {
   }
 
 
-  public void deletePscStatement(String companyNumber, String statementId) throws ResourceNotFoundException{
+  public void deletePscStatement(String contextId, String companyNumber, String statementId) throws ResourceNotFoundException{
     PscStatementDocument pscStatementDocument = getPscStatementDocument(companyNumber, statementId);
+
+    Statement statement = pscStatementDocument.getData();
+    ApiResponse<Void> apiResponse = apiClientService.invokeChsKafkaApiWithDeleteEvent(contextId, companyNumber, statementId, statement);
+
+    logger.info(String.format("ChsKafka api DELETED invoked successfully for companyNumber %s and statementId %s", companyNumber, statementId));
+    if (apiResponse == null) {
+      throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, " error response received from ChsKafkaApi");
+    }
+    HttpStatus statusCode = HttpStatus.valueOf(apiResponse.getStatusCode());
+    if (!statusCode.is2xxSuccessful()) {
+      throw new ResponseStatusException(HttpStatus.valueOf(apiResponse.getStatusCode()), " error response received from ChsKafkaApi");
+    }
+
     pscStatementRepository.delete(pscStatementDocument);
+    logger.info(String.format("Psc Statement is deleted in MongoDb with companyNumber %s and statementId %s", companyNumber, statementId));
   }
 
   private PscStatementDocument getPscStatementDocument(String companyNumber, String statementId) throws ResourceNotFoundException{
